# src/main/resources/static/swagger.yaml
swagger: '2.0'
info:
  description: This defines the API from the Blog Backend
  version: 1.0.0
  title: Swagger Blog Backend
host: localhost:8080
schemes:
  - http
paths:
...
  /users/{username}:
    get:
      summary: Get user by id
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UserResponse'
        '404':
          description: User not found
    parameters:
      - name: username
        in: path
        description: username of user
        required: true
        type: string
  /users/{username}/img:
    get:
      summary: Get profile image of given user
      produces:
        - image/png
      responses:
        '200':
          description: OK
        '204':
          description: No image for user found
        '404':
          description: User not found
    parameters:
      - name: username
        in: path
        description: username of user
        required: true
        type: string
...
definitions:
  UserResponse:
    properties:
      author:
        type: string
      vorname:
        type: string
      nachname:
        type: string
      profilBild:
        type: string
...